+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Changes in 0.1.0
+
+ * @jamesjwarren: Feature: loop (#14)
+ * @jamesjwarren: Feature: start delay (#13)
+ * @jamesjwarren: Feature: velocity button (#10)
+
 ## Changes in 0.0.4
 
  * @eopa: update documentation (no PR)